feat(dynamic-relays): move registry creator to rcm
